Iter12

+++ migrations/000002_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS metric.counters (
     id SERIAL PRIMARY KEY,
     metric_type TEXT NOT NULL DEFAULT 'counter',
     metric_name TEXT UNIQUE NOT NULL,
-    metric_value INTEGER NOT NULL,
+    metric_value BIGINT NOT NULL,
     created_at TIMESTAMPTZ NOT NULL,
     updated_at TIMESTAMPTZ NOT NULL
 );
